requirements.txt
telethon>=1.5,<2.0
pixivpy==3.5.10